requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
librosa>=0.10.0

gTTS>=2.4.0
TTS>=0.22.0
gradio>=4.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
